Add files via upload
TASK-2 :Tic-Tac-Toe Game

The game starts with two players as active members, having a one-print board on the screen that displays 9 boxes (i.e., 3×3). The first person who starts the game has to enter either X or O for any one box, followed by the other player entering the other X or O (opposite to what the first player has entered). this continues unless any of them gets a line cut diagonally or straight. And the person who founds the line is the winner of the game.